subscription

+++ app/Http/Controllers/SubscriptionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreSubscriptionRequest;
+use App\Http\Requests\UpdateSubscriptionRequest;
 use App\Models\Subscription;
 use Illuminate\Http\Request;
 
@@ -13,7 +14,31 @@ class SubscriptionController extends Controller
      */
      public function index()
     {
-      $subscriptions = Subscription::orderBy('id','DESC')->get();
+        if (auth()->check()) {
+            $user = auth()->user();
+
+            if ($user->role == 'admin') {
+
+                $subscriptions = Subscription::orderBy('id','DESC')->get();
+
+                return response()->json([
+                    'status' => true,
+                    'message' => 'Successfully fetch All subscriptions',
+                    'data' => $subscriptions,
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'You do not have the permissions to fetch All subscriptions.',
+                ], 403);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Authentication required to fetch All subscriptions.',
+            ], 401);
+        }
+    
     }
 
     /**
@@ -29,9 +54,30 @@ class SubscriptionController extends Controller
      */
     public function store(StoreSubscriptionRequest $request)
     {
-        $validatedData = $request->validated();
-        Subscription::create($validatedData);
-        return redirect()->route('Subscriptions.index')->with('success','Subscription added successfully');
+        if (auth()->check()) {
+            $user = auth()->user();
+            if ($user->role == 'admin') {
+                $validatedData = $request->validated();
+                $subscription = Subscription::create($validatedData);
+
+                return response()->json([
+                    'status' => true,
+                    'message' => 'Subscription created successfully',
+                    'data' => $subscription,
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'You do not have the permissions to create subscription.',
+                ], 403);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Authentication required to create subscription.',
+            ], 401);
+        }
+    
     }
 
     /**
@@ -39,7 +85,11 @@ class SubscriptionController extends Controller
      */
     public function show(Subscription $subscription)
     {
-        //
+        try {
+            return response()->json(['success' => true, 'message' => 'Successfully fetched the single data', 'data' => $subscription], 200);
+        } catch (\Throwable $th) {
+            return response()->json(['success' => false, 'message' => 'Failed to fetch the data', 'error' => $th->getMessage()], 500);
+        }
     }
 
     /**
@@ -47,18 +97,38 @@ class SubscriptionController extends Controller
      */
     public function edit(Subscription $subscription)
     {
-
-        return view('admin.Subscription.edit',compact('Subscription'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update( $request, Subscription $subscription)
+    public function update(UpdateSubscriptionRequest $request, Subscription $subscription)
     {
-        $validatedData = $request->validated();
-        $subscription->update($validatedData);
-        return redirect()->route('Subscriptions.index')->with('success','Subscription updated successfully');
+
+        if (auth()->check()) {
+            $user = auth()->user();
+
+            if ($user->role == 'admin') {
+                $validatedData = $request->validated();
+                $subscription->update($validatedData);
+
+                return response()->json([
+                    'status' => true,
+                    'message' => 'Subscription updated successfully',
+                    'data' => $subscription,
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'You do not have the permissions to update Subscription.',
+                ], 403);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Authentication required to update the Subscription.',
+            ], 401);
+        }
     }
 
     /**
@@ -66,7 +136,24 @@ class SubscriptionController extends Controller
      */
     public function destroy(Subscription $subscription)
     {
-        $subscription->delete();
-        return redirect()->route('Subscriptions.index')->with('success','Subscription deleted successfully');
+
+        if (auth()->check()) {
+            $user = auth()->user();
+            if ($user->role == 'admin') {
+                $deletedSubscription = $subscription;
+                $subscription->delete();
+                return response()->json(['success' => true, 'message' => 'Successfully deleted the subscription', 'data' => $deletedSubscription], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'You do not have the permissions to delete subscription.',
+                ], 403);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'Authentication required to delete subscription.',
+            ], 401);
+        }
     }
 }
